Added numeric buttons and decimal functionality in inputs

+++ app/src/main/java/com/example/calculatorapp_danielko_tryagain/MainActivity.java
@@ -17,40 +17,40 @@ public class MainActivity extends AppCompatActivity {
 
     public void add(View v){
         EditText input1 = findViewById(R.id.num1);
-        int num1 = Integer.parseInt(input1.getText().toString());
+        double num1 = Double.parseDouble(input1.getText().toString());
         EditText input2 = findViewById(R.id.num2);
-        int num2 = Integer.parseInt(input2.getText().toString());
-        int sum = num1 + num2;
+        double num2 = Double.parseDouble(input2.getText().toString());
+        double sum = num1 + num2;
         TextView answerTV = findViewById(R.id.answerTV);
         answerTV.setText("" + sum);
     }
 
     public void subtract(View v){
         EditText input1 = findViewById(R.id.num1);
-        int num1 = Integer.parseInt(input1.getText().toString());
+        double num1 = Double.parseDouble(input1.getText().toString());
         EditText input2 = findViewById(R.id.num2);
-        int num2 = Integer.parseInt(input2.getText().toString());
-        int sum = num1 - num2;
+        double num2 = Double.parseDouble(input2.getText().toString());
+        double sum = num1 - num2;
         TextView answerTV = findViewById(R.id.answerTV);
         answerTV.setText("" + sum);
     }
 
     public void multiply(View v){
         EditText input1 = findViewById(R.id.num1);
-        int num1 = Integer.parseInt(input1.getText().toString());
+        double num1 = Double.parseDouble(input1.getText().toString());
         EditText input2 = findViewById(R.id.num2);
-        int num2 = Integer.parseInt(input2.getText().toString());
-        int product = num1*num2;
+        double num2 = Double.parseDouble(input2.getText().toString());
+        double product = num1*num2;
         TextView answerTV = findViewById(R.id.answerTV);
         answerTV.setText("" + product);
     }
 
     public void divide(View v){
         EditText input1 = findViewById(R.id.num1);
-        int num1 = Integer.parseInt(input1.getText().toString());
+        double num1 = Double.parseDouble(input1.getText().toString());
         EditText input2 = findViewById(R.id.num2);
-        int num2 = Integer.parseInt(input2.getText().toString());
-        int answer = num1/num2;
+        double num2 = Double.parseDouble(input2.getText().toString());
+        double answer = num1/num2;
         TextView answerTV = findViewById(R.id.answerTV);
         answerTV.setText("" + answer);
     }
